geox-api/requirements.txt
fastapi[all]==0.78.*
snowflake-connector-python[pandas]==2.7.12
SQLAlchemy==1.4.*
celery[redis]==5.2.*
pyarrow==8.0.*
snowflake-sqlalchemy==1.4.*
geojson-pydantic==0.4.2
geopandas==0.11.*
authlib==1.0.1
starlette-authlib==0.1.11
sqlalchemy-json==0.5.0
geoalchemy2==0.12.5
psycopg2==2.9.3
httpx==0.23.0
PyJWT==2.4.0
tenacity==8.0.*
alembic==1.8.*
GitPython==3.1.*
ruamel.yaml==0.17.21
s3fs==2022.8.2
wheel
# timvt==0.7.0
git+https://github.com/developmentseed/timvt@fc12bd6a879f0ef23d427efa76fe91e47b3f8330#egg=timvt
typer==0.4.*
buildpg==0.4
hypercorn[uvloop]
alembic_utils==0.7.*
